feat: project reference to source support
Add support for project reference to source support, respecting compiler option
`disableSourceOfProjectReferenceRedirect`.

In order for this to be observed it needs to be called after the
host has been configured.

In order to simplify the logic consumer side of this package change the definition
to be about creating language service. Passing in a createLanguageService method,
which will return the language service with the proxied host as a parameter.

This allows creating a proxied host so that if the property exists on the project
object it will be used otherwise defaulting to the host. This then allows creating language
service for the project and auto import provider project.